+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Modified
 - `localhost` will be ignored when forcing the `HTTPS` protocol
+- Assets retrieval will now remove the extension from each asset, on the `get` and on the `set`
 
 ### News
 - Implemented pending features and improvements that were procrastinated for longer than necessary